requirements.txt
streamlit
camelot-py
pdfplumber
pytesseract
pandas
openpyxl
pillow
